requirements.txt
flask
SQLAlchemy
Jinja2
python-dotenv
psycopg2-binary